ref(chart): Fixes some chart problems
This fixes some documentation and less-than-ideal template usage (especially with namespaces)
that I found while giving dapr a whirl. As part of this, I have bumped the patch version
on all charts.

Some other recommendations for posterity sake that are out of scope of this PR, but
would be good to make this a more production ready chart:
- Make the name and labels for objects dynamic (generally with the fullname template)
- Consider adding redis as a dependency (that can be turned off with a condition) to
  make it easier for those getting started
- Fill out `Chart.yaml` with icons/sources/etc.

Signed-off-by: Taylor Thomas <taylor.thomas@microsoft.com>