[E2E] Initial commit of e2e test utils (#758)
This PR introduces KubeClient, and AppManager, which are the helpers to deploy app and create service, and clean up deployment and services by using k8s client-go.

K8S fake client and reactor are used for the unit-tests for these implementations.

When it comes to helloworld_test.go in this pr. The current implementation is to show how to use appmanager. The next PR will include test framework to simplify the test setup and cleanup steps. Then, helloworld_test.go will be rewritten with new test framework.