Delve Debugger vs (and just) Delve
As a windows focused developer, Delve means something very different for me, and I thought a sentence to clarify the purpose of Delve for Go on the configuration page would immediately add context and reduce confusion

As I always strive to do in any documentation - NOT the what, but the WHY is more important.  If you understand WHY the feature was developed or used, the WHAT it does and HOW starts in the right context of purpose.