Remove set-env usage (#2219)
* Remove set-env usage

https://nvd.nist.gov/vuln/detail/CVE-2020-15228 advises that setenv
and add-path are not secure, so they will be disabled very shortly on
workflows. This PR changes set-env to the recommended upgrade.

* Fix missed '##set-env' line

* Fix missed test script

* Fix missed Python script

* Use file.write instead of print(..., file=x)

* Use latest setup-go, which removed set-env

* Remove multiline string usage

Apparently environment variable names can have whitespace. This commit
removes the multiline string usage to avoid having whitespace in the
variable names we set (thus setting the wrong env var).

Co-authored-by: Mukundan Sundararajan <musundar@microsoft.com>