Update setup-dapr-development-env.md
If you set GOPATH to C:\go as displayed you can't install Delve. You will get the following error:
```
C:\go is a GOROOT, not a GOPATH. For more details see: 'go help gopath'
```
After installing go with the MSI from the go site, GOPATH is already set to %USERPROFILE%\go which is what the `go help gopath` says it should be.