Use Aborted error code for ETag mismatch
Related to #2619

I started a discussion after https://github.com/dapr/dapr/pull/2647 was
merged because I think there's a better fit for an ETag mismatch than
AlreadyExists.

The gRPC docs have specific guidance that Aborted is the best status for
'when a client-specified test-and-set fails', which is this case.