Avoid push of Helm chart when chart version wasn't updated
This changes the workflow to check if the intended change actually
contains a file with ${daprVersion} in the name. In the case that Helm
chart version wasn't updated before tagging, we will only see changes
to the old Helm chart version and will not push, but instead failing
with an error.