Add reentrancy to actor locking (#3145)
* Add configuration for actor reentrancy

This adds the necessary fields to enable reentrancy for actors
but does not actually implement any of the reentrant behavior.

https://github.com/dapr/dapr/issues/3049

* Add reentrancy to actor locking

Actors that are specified as reentrant are now capable of bypassing
their lock when a call that is servicing the same request is found.
The request ID (or reentrancy ID) used here is generated on the
first call to the reentrant actor and is passed through all
remaining calls.

https://github.com/dapr/dapr/issues/3078

* Add e2e test for actor reentrancy

Co-authored-by: Artur Souza <artursouza.ms@outlook.com>