Lock placement table updates at operation level
Under specific circumstances, the actor placement table can get out
of sync and not have new disseminations queued. This can happen
when a dissemination begins followed by an update which finishes
before the dissemination. This puts the placement table into a state
where it does not know there are more updates when one has not been
sent.

https://github.com/dapr/dapr/issues/3023