Remove error for Actor Timer DueTime in the past (#4518) (#4519)
The DueTime of an Actor Timer signifies the first time it should
fire. An omitted or 0 DueTime means fire immediately. A check was
introduced that forced a failure if the DueTime was in the past.
However, when the time was 0, it was seen as in the past as the
evaluation logic used multiple time.Now instances.

Regardless, the semantic behavior of this did not make sense as
a DueTime in the past should still just mean fire immediately. This
commit removes that check allowing for this behavior. This is also
how Actor Reminders work.

https://github.com/dapr/dapr/issues/4514

Signed-off-by: Hal Spang <halspang@microsoft.com>