Add built-in retry for finding actor in placement (#4803)
This change adds a built-in policy for finding an actor in the
placement table. This policy is protected and cannot be touched by
customers. This is because this type of retry should be hidden from
general users as we have the most information on how the placement
service works/values that should be used for the policy.

Signed-off-by: Hal Spang <halspang@microsoft.com>