Fix resiliency logic in gRPC proxying (#4829)
In the case of gRPC proxying, we need to record all the calls that
have occurred and replay them if we need to perform retries. The
array of calls was being shared between goroutines which was causing
different calls to share frames. This causes requests to exhibit
incorrect behavior/fail. The fix for this is to isolate the call
array.

Additionally, this commit also fixes an issue with headers being
sent during stream calls. When a stream fails mid-stream, it
attempted to send the initial headers again. This hung forever as
the listener was no longer expecting headers.

https://github.com/dapr/dapr/issues/4537

Signed-off-by: Hal Spang <halspang@microsoft.com>
Signed-off-by: Bernd Verst <4535280+berndverst@users.noreply.github.com>
Signed-off-by: Addison Juarez <adjuarez@microsoft.com>