Fix missing components in Metadata API result.
Dapr's GRPC Metadata API was never returning the list
of registered component. That happened because the API
object had a `components` attribute that was never assigned
during API construction - it received a `getComponentsFn`
but that one was also never invoked.

This PR addressed this by aligning its implementation with
its HTTP counterpart - removing the `components` attribute
and using the `getComponentsFn` provided to the constructor
instead.

Fixes #5051

Signed-off-by: Tiago Alves Macambira <tmacam@burocrata.org>