Sanitize gRPC headers before sending to bindings
The gRPC headers such as ':authority' were not being sanitized before
being sent to the various bindings. As a result, calls had the
possibility of failing. This was found in the certification tests.

This commit sanitizes the headers first so they may be passed without
interfering. This should not impact any gRPC functionality as these
headers were being ignored in the past.

Signed-off-by: Hal Spang <halspang@microsoft.com>