Full support for `automountServiceAccountToken: false`
The injector is changed to create a new projected volume with a service account token with an audience specific for Dapr's Sentry service. Also changes the default value for --token-audience in Sentry to be the one the injector uses.

This effectively means that daprd does not require a service account token with access to the Kubernetes APIs anymore.

Completes the work started on #5359

BREAKING: This requires Kubernetes 1.22 or higher for projected token volumes to be stable
Signed-off-by: ItalyPaleAle <43508+ItalyPaleAle@users.noreply.github.com>