Full support for `automountServiceAccountToken: false` - use bound service account tokens (#5371)
* Full support for `automountServiceAccountToken: false`

The injector is changed to create a new projected volume with a service account token with an audience specific for Dapr's Sentry service. Also changes the default value for --token-audience in Sentry to be the one the injector uses.

This effectively means that daprd does not require a service account token with access to the Kubernetes APIs anymore.

Completes the work started on #5359

BREAKING: This requires Kubernetes 1.22 or higher for projected token volumes to be stable
Signed-off-by: ItalyPaleAle <43508+ItalyPaleAle@users.noreply.github.com>

* 💄

Signed-off-by: ItalyPaleAle <43508+ItalyPaleAle@users.noreply.github.com>

* Fixes for tests

Signed-off-by: ItalyPaleAle <43508+ItalyPaleAle@users.noreply.github.com>

* Smoother upgrade path for control plane services

When Sentry isn't started with `--token-audience`, allow falling back to the nil token audience (ie. the default one for the Kubernetes API server) unless the NoDefaultTokenAudience feature is enabled

Signed-off-by: ItalyPaleAle <43508+ItalyPaleAle@users.noreply.github.com>

* Remove feature flag

Signed-off-by: ItalyPaleAle <43508+ItalyPaleAle@users.noreply.github.com>

* Mark token volume as read-only

Signed-off-by: ItalyPaleAle <43508+ItalyPaleAle@users.noreply.github.com>

* Move bound token to another volume

This allows _also_ mounting the Kuberentes API token if apps need that

Fixes support for the K8s secret store component

Signed-off-by: ItalyPaleAle <43508+ItalyPaleAle@users.noreply.github.com>

Signed-off-by: ItalyPaleAle <43508+ItalyPaleAle@users.noreply.github.com>
Co-authored-by: Yaron Schneider <schneider.yaron@live.com>
Co-authored-by: Artur Souza <artursouza.ms@outlook.com>