Remove pluggable components from E2E tests.
These tests currently depend on the Pluggable Components Go SDK, which
is a dapr-sandbox project (`dapr-sandbox/components-go-sdk`).

While valuable to tests the pluggable component interface, Dapr core
repositories (such as `dapr/dapr`) should not have dependencies on
projects in dapr-sandbox.

Ideally, we should have these E2E tests coded without depending on
pluggable components SDKs. Or at the very least, depend on pinned
version of them _once they graduate_. But that's for the medium/long
term.

I am removing these tests for the time being to address this
dapr-sandbox dependency in the short term. We can revert this if
required.

Signed-off-by: Tiago Alves Macambira <tmacam@burocrata.org>