Allow enabling preview features at build-time
Added the `ENABLED_FEATURES` env var to the Makefile to define a (comma-separated) list of features that are always enabled, regardless of what's in the Configuration spec.

The `Resiliency` feature was added to the list of always-enabled features for now (replacing the previous "hack" to have it always enabled - see #5523).

Signed-off-by: ItalyPaleAle <43508+ItalyPaleAle@users.noreply.github.com>