Fixes metric grouping for CPU usage graphs. (#5525)
"Total CPU Usage" graph is displaying `container_cpu_usage_seconds_total`
but it is not grouping it by application and is instead using the
`pod` field as its discriminating field.

Each application is re-deployed daily and receives a new pod-id. As
this metric graph uses the `pod` id in its legend, the same
application ends up in represented as a series of disconnected
metrics.

This PR fixes the metric to grouping distinct "pods" under the same
by "application id" using some metric-math. This in turn will allow
us to observe how a given application behaves over time.

Fixes #5524

Signed-off-by: Tiago Alves Macambira <tmacam@burocrata.org>

Signed-off-by: Tiago Alves Macambira <tmacam@burocrata.org>
Co-authored-by: Loong Dai <long.dai@intel.com>
Co-authored-by: Artur Souza <asouza.pro@gmail.com>