Resiliency: do not retry on HTTP service invocation for 2xx status codes
When performing HTTP service invocation, only responses with a 200 status code were considered successful, and other 2xx were treated as failures. Because of that, even thought he user would ultimately get the response with a 2xx status code as if "nothing happened", the request would be retried if the resiliency policy had retries on.

This only applied to HTTP APIs; gRPC ones were already working as expected (when calling HTTP endpoints)

Signed-off-by: ItalyPaleAle <43508+ItalyPaleAle@users.noreply.github.com>