Fix actor reminder locks to prevent race conditions with evaluations
Currently, a channel (`evaluationChan`) used as mutex prevents reminders to be saved or deleted while reminder evaluation (which happens every time the placement table changes) is in progress. However, nothing was preventing reminder evaluation to begin while a save/delete operation on reminders was in progress.

This changes the use of locks in the package so evaluation can't start while a reminder is being created/deleted.

Signed-off-by: ItalyPaleAle <43508+ItalyPaleAle@users.noreply.github.com>