Reduce duration of workflow perf tests
Currently, the tests take close to 2hrs to complete. This reduces the amount of effort of each test, and should reduce the duration of the test by a noticeable amount (although the tests will still take much longer than any other perf test). This will be without sacrificing the quality of the data.

Signed-off-by: ItalyPaleAle <43508+ItalyPaleAle@users.noreply.github.com>