[1.12] Fix sentry rejecting legacy request IDs of length larger than 64 (#7082)
* Fix sentry rejecting legacy request IDs of length larger than 64

Fixes https://github.com/dapr/dapr/issues/7078

Fixes an issue whereby sentry would reject SignCertificate requests
that have an identifier which was longer than 64 characters. Because
sentry still accepts the "legacy" identifier format of
`<namespace>:<service account>` (which v1.12 is currently using),
Daprd's which are in a namespace and/or using a Service Account with a
sufficiently long name will erroneously be rejected.

PR updates the Kubernetes sentry validator so that it will augment the
request identifier to the app ID before evaluating whether the
identifier is over 64 characters long.

Adds sentry Kubernetes validator integration tests to cover this
scenario.

Signed-off-by: joshvanl <me@joshvanl.dev>

* Adds `v1.12.1.md` release notes

Signed-off-by: joshvanl <me@joshvanl.dev>

---------

Signed-off-by: joshvanl <me@joshvanl.dev>
Signed-off-by: Cassandra Coyle <cassie@diagrid.io>