Adds `--trust-anchors-file` to Daprd
Part of the joshvanl proposal trust distribution
https://github.com/dapr/proposals/pull/42

PR updates kit to use the kit/fswatcher that uses the events/batcher to
de-dupe file events.

Updates daprd with a new `--trust-anchors-file` flag and environment
variable `DAPR_TRUST_ANCHORS_FILE` to allow users to specify a reference
to a file containing the trust anchors bundle. This new flag supersedes
the environment variable, which both supersede the existing static
`DAPR_TRUST_ANCHORS` environment variable.

The trust anchor in security used by daprd will be updated to match the
contents of the referenced file when the file changes. An environment
variable is added so that backwards compatibility can be achieved when
injector is updated to populate file reference, whilst still supporting
patching older Daprds.

Integration tests have been added to validate the security CA bundle is
updated when the file changes for both daprd when using the env var or
flag, as well as sentry.

Signed-off-by: joshvanl <me@joshvanl.dev>