Integration: Only write process logs on failure (#7074)
* Integration: Only write logs on failure

Signed-off-by: joshvanl <me@joshvanl.dev>

PR updates the integration test io writer to by default, only write the
exec process logs on test failure. This is in an effort to reduce the
amount of IO done by the test runners. If a test fails, the processes
logs will be output at the end. Logs can be enabled for all tests by
setting the `DAPR_INTEGRATION_LOGS` environment variable to `true`.

* Updates tests/docs/writing-integration-test.md to remove unneeded new
lines

Signed-off-by: joshvanl <me@joshvanl.dev>

---------

Signed-off-by: joshvanl <me@joshvanl.dev>
Co-authored-by: Dapr Bot <56698301+dapr-bot@users.noreply.github.com>
Signed-off-by: Cassandra Coyle <cassie@diagrid.io>