Fix dissemination of placement tables
Fixes (hopefully) #7193

Fixes two issues in the process of disseminating placement tables.

First, dissemination should always have a background context, and not be tied to the context of a specific stream. Otherwise, if the stream is disconnected before the dissemination, all connected hosts may be left in a broken state.

Second, the dissemination process was meant to follow a 3-step process: acquire a lock, disseminate the tables, then release the lock. The current implementation performed all 3 operations in parallel, effectively not waiting for all nodes to be locked before disseminating the tables. This means that the nodes could potentially have inconsistent tables.

Signed-off-by: ItalyPaleAle <43508+ItalyPaleAle@users.noreply.github.com>
Signed-off-by: Cassandra Coyle <cassie@diagrid.io>