Reduce duration of workflow perf tests (#6999)
* Reduce duration of workflow perf tests

Currently, the tests take close to 2hrs to complete. This reduces the amount of effort of each test, and should reduce the duration of the test by a noticeable amount (although the tests will still take much longer than any other perf test). This will be without sacrificing the quality of the data.

Signed-off-by: ItalyPaleAle <43508+ItalyPaleAle@users.noreply.github.com>

* Improve workflow perf test app

Signed-off-by: ItalyPaleAle <43508+ItalyPaleAle@users.noreply.github.com>

* Reduce duration of workflow perf tests

Currently, the tests take close to 2hrs to complete. This reduces the amount of effort of each test, and should reduce the duration of the test by a noticeable amount (although the tests will still take much longer than any other perf test). This will be without sacrificing the quality of the data.

Signed-off-by: ItalyPaleAle <43508+ItalyPaleAle@users.noreply.github.com>

* Improve workflow perf test app

Signed-off-by: ItalyPaleAle <43508+ItalyPaleAle@users.noreply.github.com>

* Changed parallels

Signed-off-by: ItalyPaleAle <43508+ItalyPaleAle@users.noreply.github.com>

* Tiny fix

Signed-off-by: ItalyPaleAle <43508+ItalyPaleAle@users.noreply.github.com>

---------

Signed-off-by: ItalyPaleAle <43508+ItalyPaleAle@users.noreply.github.com>
Co-authored-by: Mukundan Sundararajan <65565396+mukundansundar@users.noreply.github.com>
Co-authored-by: Dapr Bot <56698301+dapr-bot@users.noreply.github.com>
Signed-off-by: Cassandra Coyle <cassie@diagrid.io>