Updates controller-runtime, remove WriteSVIDToDir
PR updates go modules, specifically controller-runtime to v0.16.3 so
that the webhook server can be configured to use TLS options. This
removes the need for the operator to write the SVID to disk and maintain
it in memory.

PR removes the WriteSVIDToDir option from security.

Operator cache has been updated to use new controller-runtime types.

Signed-off-by: joshvanl <me@joshvanl.dev>