Integration: Reduce test output noise when focusing tests (#7216)
In integration tests, ee currently use `t.Skip` when skipping tests
that don't match on the given focus (`-focus`). This results in noisy
test output which makes it difficult to parse results on tests which the
developer cares about.

PR updates the integration test framework to not register focused tests
at all and instead only print a log line, greatly reducing the noise
when running tests with `-focus`.

Signed-off-by: joshvanl <me@joshvanl.dev>
Signed-off-by: Cassandra Coyle <cassie@diagrid.io>