Split up runtime resource handler into processor manager (#7086)
* Split up runtime resource handler into processor manager

Signed-off-by: joshvanl <me@joshvanl.dev>

PR splits out the runtime authorizer and resource processor into a
dedicated authorizer package, and a manager in the processor package.
This de-couples the main runtime package from the processor so that hot
reloading (not implemented here) can also make use of the processor
manager.

Components are no longer indexed in the ComponentStore by both their name
_and_ type. This is because the hot reloader would not be able to
determine "which" Component it should delete on delete events. No two
Components with the same type but different names can co-exist anyway.

* Linting

Signed-off-by: joshvanl <me@joshvanl.dev>

* Populate operator client in secret processor

Signed-off-by: joshvanl <me@joshvanl.dev>

* Adds authorizer_test.go

Signed-off-by: joshvanl <me@joshvanl.dev>

---------

Signed-off-by: joshvanl <me@joshvanl.dev>
Co-authored-by: Dapr Bot <56698301+dapr-bot@users.noreply.github.com>
Signed-off-by: Cassandra Coyle <cassie@diagrid.io>