Helm: Add option to deploy Placement with zero replicas
The `dapr_placement.scaleZero` option passed to Helm causes the Placement StatefulSet to be deployed but with 0 replicas, regardless of the values of `dapr_placement.ha` or `global.ha.enabled`.

This is useful in a few niche scenarios, where it's desirable to have the StatefulSet created, but in a "paused" (i.e. scaled to zero) state, to be able to scale it up later.

> This is different from setting `global.actors.enabled=false`, which causes the Placement StatefulSet, and its related Service etc, to not be deployed at all

Signed-off-by: ItalyPaleAle <43508+ItalyPaleAle@users.noreply.github.com>