Chore: update artifact actions
The `master` tag was pointing to a very old build that sometimes failed (the "master" branch is now "main"). More appropriately pointing to a version now. Also, changed to avoid compressing artifacts that are already compressed.

In the process, also updated other actions including setup-go, which comes with its own caching for Go deps so we don't need the custom cache tasks

Signed-off-by: ItalyPaleAle <43508+ItalyPaleAle@users.noreply.github.com>