Updates controller-runtime, remove WriteSVIDToDir (#7214)
* Updates controller-runtime, remove WriteSVIDToDir

PR updates go modules, specifically controller-runtime to v0.16.3 so
that the webhook server can be configured to use TLS options. This
removes the need for the operator to write the SVID to disk and maintain
it in memory.

PR removes the WriteSVIDToDir option from security.

Operator cache has been updated to use new controller-runtime types.

Signed-off-by: joshvanl <me@joshvanl.dev>

* Update deprecated funcs

Signed-off-by: joshvanl <me@joshvanl.dev>

* Replace deprecated function in test/runner

Signed-off-by: joshvanl <me@joshvanl.dev>

* Fix Kubernetes integration process

Signed-off-by: joshvanl <me@joshvanl.dev>

* Include int framework operator v2alpha1 API endpoint with subscription
v2

Signed-off-by: joshvanl <me@joshvanl.dev>

---------

Signed-off-by: joshvanl <me@joshvanl.dev>
Co-authored-by: Yaron Schneider <schneider.yaron@live.com>
Co-authored-by: Dapr Bot <56698301+dapr-bot@users.noreply.github.com>