Workflow backends: exit error on hot reload
Updates the hot reloading reconciler so that Daprd will exit error when
a workflowbackend Component is hot reloaded. This is chosen because
today, the actors and workflow subsystems are not written with any
closing or dynamic support. Doing so will cause panics/corruption in its
current state.

Exiting error is the safest option as this ensures consistency across a
replica set and ensures there is no surprise for the user that
behaviour does not match given configuration.

Adds integration tests for daprd ensuring workflow backends can be
loaded on boot.

Adds tests to ensure daprd will exist error on a workflow backend being
hot reloaded.

We should do the same for actor state store and the actor subsystem.

Signed-off-by: joshvanl <me@joshvanl.dev>