Actor State: Ignore Components which hot reload the actor state (#7441)
* Exit error if actor state store hot reloaded

Updates the hot reloading reconciler so that Daprd will exit error when
a actor state store enabled Component is hot reloaded. This is chosen
because today, the actors subsystem is not written with any closing or
dynamic support. Doing so will cause panics/corruption in its current
state.

Exiting error is the safest option as this ensures consistency across a
replica set and ensures there is no surprise for the user that behaviour
does not match given configuration.

See also https://github.com/dapr/dapr/pull/7433

Signed-off-by: joshvanl <me@joshvanl.dev>

* Remove expected workflow.dapr component from actorastate tests

Signed-off-by: joshvanl <me@joshvanl.dev>

* Increase assert Eventually timeout

Signed-off-by: joshvanl <me@joshvanl.dev>

* Only log error and don't exit error when reocnciling actor state store

Signed-off-by: joshvanl <me@joshvanl.dev>

---------

Signed-off-by: joshvanl <me@joshvanl.dev>
Signed-off-by: Artur Souza <asouza.pro@gmail.com>
Co-authored-by: Artur Souza <asouza.pro@gmail.com>