Workflow backends: ignore on hot reloading (#7433)
* Workflow backends: exit error on hot reload

Updates the hot reloading reconciler so that Daprd will exit error when
a workflowbackend Component is hot reloaded. This is chosen because
today, the actors and workflow subsystems are not written with any
closing or dynamic support. Doing so will cause panics/corruption in its
current state.

Exiting error is the safest option as this ensures consistency across a
replica set and ensures there is no surprise for the user that
behaviour does not match given configuration.

Adds integration tests for daprd ensuring workflow backends can be
loaded on boot.

Adds tests to ensure daprd will exist error on a workflow backend being
hot reloaded.

We should do the same for actor state store and the actor subsystem.

Signed-off-by: joshvanl <me@joshvanl.dev>

* Review loop

Signed-off-by: joshvanl <me@joshvanl.dev>

* Update workflowbackend int tests for removed workflow comoponent
registered

Signed-off-by: joshvanl <me@joshvanl.dev>

* Fix namespacing in workflowbackend integration test

Signed-off-by: joshvanl <me@joshvanl.dev>

* Changes hot reloading to only log an error when workflow backend is
reconciled

Signed-off-by: joshvanl <me@joshvanl.dev>

* Fix bad merge conflict resolution.

Signed-off-by: Artur Souza <asouza.pro@gmail.com>

---------

Signed-off-by: joshvanl <me@joshvanl.dev>
Signed-off-by: Artur Souza <asouza.pro@gmail.com>
Co-authored-by: Dapr Bot <56698301+dapr-bot@users.noreply.github.com>
Co-authored-by: Artur Souza <asouza.pro@gmail.com>
Signed-off-by: Cassandra Coyle <cassie@diagrid.io>