Workflow delete: Remove `non_recursive` with hardcoded always enabled `recursive` (#7498)
* Workflow delete: Change double negative `non_recursive` to `recursive`

The workflow terminate and purge APIs currently take a `non_recursive`
parameter to indicate whether to delete the workflow and its children.
This is a double negative and is confusing for users, needing a few more
brain cycles to understand the field. We should avoid double negatives
in APIs. This PR changes the parameter to `recursive` to make it more
clear.

`recursive` defaults to `true`.

Signed-off-by: joshvanl <me@joshvanl.dev>

* Linting

Signed-off-by: joshvanl <me@joshvanl.dev>

* Linting

Signed-off-by: joshvanl <me@joshvanl.dev>

* Completely remove option for recursive workflow deletion

Signed-off-by: joshvanl <me@joshvanl.dev>

* Linting

Signed-off-by: joshvanl <me@joshvanl.dev>

* Linting

Signed-off-by: joshvanl <me@joshvanl.dev>

---------

Signed-off-by: joshvanl <me@joshvanl.dev>
Co-authored-by: Mukundan Sundararajan <65565396+mukundansundar@users.noreply.github.com>