Healthz overhaul
Currently, the healthz endpoint on Dapr services do not represent the
actual health of the service as they are mostly only spinning up a HTTP
server and immediately returning a 200 OK. This results in services
receiving traffic before they are ready to actually serve requests. This
issue manifests in integration tests where, for example, no metrics are
available on the metrics endpoint even though the service reports as
healthy and therefore fails the test.

This PR overhauls the healthz package to allow registering health
targets during the service `New` process chain. Each module marks their
target as healthy and the healthz server reports the overall health of
the service. Although not perfect, service healthz endpoints should now
be more representative of the actual health of the program.

Signed-off-by: joshvanl <me@joshvanl.dev>