Refactor api/errors into builders
Refactors api/errors into functional builders. The current API for
api/errors is heavily string concatenation based with long parameter
arguments. This is fragile/error prone to the developer and difficult to
understand for the consumer.

This change should make it clearer to the reader/consumer what the
intended errors are and how to use them. Moving implementation details
of the error protocol codes into the error package also helps to clean
up API implementation code.

Signed-off-by: joshvanl <me@joshvanl.dev>