[1.13] cherry pick: Fix daprd hotreload operator stream re-establishment causing tight loop (#7611)
* Fix daprd hotreload operator stream re-establishment causing tight loop (#7609)

When the HotReload feature is enabled in Kubernetes mode, daprd streams
Component changes from the operator. In the event this stream failing
(for example when the operator pod is restarted), daprd will attempt to
re-establish the stream.

Due to a bug in the stream re-establishment logic, daprd would tight
loop attempting to re-establish the stream. This results in component
updates not being sent to daprd, as well as causing the operator to
catastrophically print hundreds of connection logs a second. This
completely locks up the operator process, and will produce GBs of logs
unless _all daprds are restarted_.

PR fixes the tight loop by correctly breaking out of retries after a
successful re-establishment. Also ensures daprd will re-reconcile all
Components in the event of a stream re-establishment, catching any
Components which might have been deleted while the stream was down.

Adds integration tests operator stream re-establishment.

Signed-off-by: joshvanl <me@joshvanl.dev>

* Adds v1.13.1.md release notes

Signed-off-by: joshvanl <me@joshvanl.dev>

---------

Signed-off-by: joshvanl <me@joshvanl.dev>