Fixed Actor metadata is stored under 'actors||actortype' partitionkey
Changed that the Actor metadata is stored under correct partitionkey
via 'constructCompositionKey' instead of 'calculatePartitionKey' which
is used for the reminders partitions.

This results in that the Actor metadata can be found and updated when
a reminder gets stored or deleted, otherwise this will create a new
metadata record with incorrect metadata configuration like 'partitionCount'.

Issue: #3380 (components-contrib)

Signed-off-by: Mark van den Bosch <markvandenbosch@gmail.com>