Fix incorrect content-length being sent to HTTP published message (#7659)
* Fix incorrect content-length being sent to HTTP published message

PR #7537 reverse revered the change which removed the content-length
from being set on HTTP headers on sent messages. We still need to remove
the content-length from messages from pubsub subscriptions as the pubsub
may report a message with a content-length which does not actually
match the size of the delivered message to the app.

content-length is only removed on HTTP published messages.

Change should be backported to 1.13

Signed-off-by: joshvanl <me@joshvanl.dev>

* Adds grpc app subscriber to content-length tests

Signed-off-by: joshvanl <me@joshvanl.dev>

* Adds tests for content-length gRPC subscribed app

Signed-off-by: joshvanl <me@joshvanl.dev>

* framework/socket: skip if test is windows

Signed-off-by: joshvanl <me@joshvanl.dev>

* Fix socket runtime GOOS import

Signed-off-by: joshvanl <me@joshvanl.dev>

---------

Signed-off-by: joshvanl <me@joshvanl.dev>
Signed-off-by: Joey Freeland <joey@free.land>