Operator: Component server side scope filtering (#7615)
* Operator: Component server side scope filtering

Today, daprd's will receive Component from the operator, regardless of
whether they are scoped for that Component or not. This means that
clients do receive components (including its associated secrets) that
they are not scoped for.

Updates Operator API ComponentUpdate to perform service side Component
Scope filtering based on the authenticated client App ID. When a
Component is de-scoped, daprd will receive a DELETE for the previous
Component manifest. When a Component is scoped-in, daprd will receive an
CREATE for the new Component manifest.

Updates Operator API ListComponents to perform service side Component
Scope filtering based on the authenticated client App ID.

Uses updated events batcher which ensures queue items are sent in order.

Signed-off-by: joshvanl <me@joshvanl.dev>

* Add 2 daprds for operator informer tests

Signed-off-by: joshvanl <me@joshvanl.dev>

* Fix control plane trust domain

Signed-off-by: joshvanl <me@joshvanl.dev>

* Update github.com/dapr/kit to master

Signed-off-by: joshvanl <me@joshvanl.dev>

* go mod tidy

Signed-off-by: joshvanl <me@joshvanl.dev>

* Fix incorrect pointer manifest compare

Signed-off-by: joshvanl <me@joshvanl.dev>

---------

Signed-off-by: joshvanl <me@joshvanl.dev>
Co-authored-by: Dapr Bot <56698301+dapr-bot@users.noreply.github.com>
Co-authored-by: Yaron Schneider <schneider.yaron@live.com>