Healthz overhaul (#7518)
* Healthz overhaul

Currently, the healthz endpoint on Dapr services do not represent the
actual health of the service as they are mostly only spinning up a HTTP
server and immediately returning a 200 OK. This results in services
receiving traffic before they are ready to actually serve requests. This
issue manifests in integration tests where, for example, no metrics are
available on the metrics endpoint even though the service reports as
healthy and therefore fails the test.

This PR overhauls the healthz package to allow registering health
targets during the service `New` process chain. Each module marks their
target as healthy and the healthz server reports the overall health of
the service. Although not perfect, service healthz endpoints should now
be more representative of the actual health of the program.

Signed-off-by: joshvanl <me@joshvanl.dev>

* Fix healthz server in injector

Signed-off-by: joshvanl <me@joshvanl.dev>

* Adds daprd healthz mtls check

Signed-off-by: joshvanl <me@joshvanl.dev>

* Fix healthz/daprd error string match for windows

Signed-off-by: joshvanl <me@joshvanl.dev>

* Fix reconciler unit test

Signed-off-by: joshvanl <me@joshvanl.dev>

* Name operator healthz target set

Signed-off-by: joshvanl <me@joshvanl.dev>

* Fix correct health var name for sub update informer

Signed-off-by: joshvanl <me@joshvanl.dev>

* Fix rebased unit tests

Signed-off-by: joshvanl <me@joshvanl.dev>

* Linting

Signed-off-by: joshvanl <me@joshvanl.dev>

* Return json marshal of placement state table for getter

Signed-off-by: joshvanl <me@joshvanl.dev>

* Linting

Signed-off-by: joshvanl <me@joshvanl.dev>

* feat: add new healthz to scheduler

Signed-off-by: mikeee <hey@mike.ee>

* chore: remove existing health handler

Signed-off-by: mikeee <hey@mike.ee>

* fix: add missing healthz

Signed-off-by: mikeee <hey@mike.ee>

* fix: init healthz server

Signed-off-by: mikeee <hey@mike.ee>

* Pass Healthz to the Scheduler Options in integration test

Signed-off-by: joshvanl <me@joshvanl.dev>

---------

Signed-off-by: joshvanl <me@joshvanl.dev>
Signed-off-by: mikeee <hey@mike.ee>
Co-authored-by: Dapr Bot <56698301+dapr-bot@users.noreply.github.com>
Co-authored-by: mikeee <hey@mike.ee>