Scheduler: Fix HA mode in Kubernetes
Fix scheduler HA mode in Kubernetes by correctly injecting the entire
scheduler address list to dapds. This is done by getting the Scheduler
StatefulSet, and building the address list via the number of replicas.

The injector now has permissions to Get StatefulSets in the control
plane namespace. This should be a safe permission to give.

Signed-off-by: joshvanl <me@joshvanl.dev>